Move to YAML + pandoc
Rework the existing CV into a YAML file containing all metadata
associated with the CV and a LaTeX template.  This both makes the
maintenance of the CV much easier as all that needs to be updated is the
YAML file and enables better usage of this CV template by others.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>